+++ FirebaseVertexAI/CHANGELOG.md
@@ -1,4 +1,11 @@
 # Unreleased
+- [added] **Public Preview**: Added support for specifying response modalities
+  in `GenerationConfig`. This includes **public experimental** support for image
+  generation using Gemini 2.0 Flash (`gemini-2.0-flash-exp`). (#14658)
+  <br /><br />
+  Note: This feature is in Public Preview and relies on experimental models,
+  which means that it is not subject to any SLA or deprecation policy and could
+  change in backwards-incompatible ways.
 - [added] Added support for specifying the minimum and maximum number of items
   (`minItems` / `maxItems`) to generate in an array `Schema`. (#14671)